pyproject.toml
[project]
name = "ctc-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.78.0",
    "openai>=2.24.0",
    "openai-agents>=0.10.2",
    "google-genai>=1.0.0",
    "google-adk>=1.0.0",
]

[dependency-groups]
dev = [
    "notebook>=7.5.4",
    "ruff>=0.15.0",
]
